:recycle: clean up cmd structure more simple
